+++ src/main/java/model/player/PlayerName.java
@@ -1,4 +1,4 @@
-package model.playername;
+package model.player;
 
 import java.util.Objects;
