+++ src/main/java/model/player/Players.java
@@ -1,4 +1,4 @@
-package model.playername;
+package model.player;
 
 import java.util.List;
 import java.util.Objects;
