+++ src/test/java/model/player/PlayerNameTest.java
@@ -1,6 +1,5 @@
-package model.playername;
+package model.player;
 
-import model.playername.PlayerName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
