+++ src/test/java/model/player/PlayersTest.java
@@ -1,4 +1,4 @@
-package model.playername;
+package model.player;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
